add 'ephemeral' / testing deploy stage
deploys a version of the site on an 'ephemeral' hypernode, which is an on-the-fly created copy of the $APP_NAME Hypernode. the `$testingStage->addEphemeralServer($APP_NAME)` step uses the [hypernode-api-php](https://github.com/ByteInternet/hypernode-api-php) client library for the Hypernode API to create that disposable environment.

the test environment can be deployed like:
```
$ docker run --rm -it --env HYPERNODE_API_TOKEN=$HYPERNODE_API_TOKEN --env SSH_PRIVATE_KEY="$(cat ~/.ssh/yourdeploykey | base64)" -v ${PWD}:/build quay.io/hypernode/deploy:latest hypernode-deploy deploy test -vvv
```