add example of deploying to hypernode-docker
We can also deploy to a Hypernode Docker instance. To do that you go to https://github.com/byteinternet/hypernode-docker, make sure you have an instance running by for example doing:
```
$ sudo docker run -P docker.hypernode.com/byteinternet/hypernode-buster-docker-php80-mysql57:latest
```
and then noting the IP address (in my case 172.17.0.2). You then need to make sure your deploykey public key is added to the /data/web/.ssh/authorized_keys file. Then you should be able to deploy to the container as if it was a 'real' hypernode.

Keep in mind that the hypernode-docker is not a real VM, it's just a fat container. This means that there won't be an init system (no systemd) so the processes are running in SCREENs. Also obviously you can not use some of the hypernode command-line functionality that depends on the Hypernode API (it's not a server managed by the Hypernode automation, just a local container running on your PC).

Note that Hypernode Deploy wouldn't be able to start a hypernode-docker instance for you on your host (like how it is able to start Brancher nodes on the fly) because Hypernode Deploy itself is also running in a Docker container on your host. You need to have a Hypernode Docker instance on your host running before you are able to deploy to it (on 172.17.0.2 for example, if you are on windows or mac you might need to figure out some port forwarding and use localhost).